+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- (qt-backend) Gradient with `objectBoundingBox` rendering.
+- (qt-backend) Text bounding box detection during the rendering.
 
 ## [0.2.0] - 2018-04-24
 ### Added